+++ client/app/app.component.html
@@ -1,1 +1,1 @@
-<router-outlet></router-outlet>
+<app-landing-page></app-landing-page>
