+++ client/app/auth-module/auth.module.ts
@@ -1,3 +1,5 @@
+import { authRoutes } from './auth.routes';
+import { AuthComponent } from './auth/auth.component';
 import { RouterModule } from '@angular/router';
 import { SignupComponent } from './signup/signup.component';
 import { LoginComponent } from './login/login.component';
@@ -9,12 +11,12 @@ import { NgModule } from '@angular/core';
 @NgModule({
     declarations: [
         LoginComponent,
-        SignupComponent,
+        SignupComponent
     ],
     imports: [
         CommonModule,
         ReactiveFormsModule,
-        RouterModule
+        RouterModule.forChild(authRoutes)
     ],
     providers: [
