--- client/app/auth-nav/auth.component.html
@@ -1,11 +1,0 @@
-
-  <nav class="card-header" role="navigation">
-      <ul class="nav nav-pills justify-content-center navbar-inverse">
-        <li class="nav-item" ><a routerLinkActive="active-link" class="btn" [routerLink]="['/login']">Log in</a></li>
-        <li class="nav-item" ><a routerLinkActive="active-link" class="nav-link" [routerLink]="['/signup']">Sign Up</a></li>
-      </ul>
-    </nav>
-
-
-            
-
