--- client/app/auth.module/login/login.component.html
@@ -1,35 +1,0 @@
-<div class="container">
-  <div class="row align-items-center">
-    <div class="col-12">
-      <form [formGroup]="loginForm" (ngSubmit)="login()">
-        <div class="form-group row justify-content-center">
-            
-          <div class="col-12 col-sm-5">
-              <div class="input-group">
-                  <div class="input-group-addon"><span class="fa fa-envelope-o"></span></div>
-            <input type="text" class="form-control" id="email" formControlName="email"
-            placeholder="Email">
-              </div>
-          </div>
-        </div>
-          <div class="form-group row justify-content-center">
-           
-            <div class="col-12 col-sm-5">
-                <div class="input-group">
-                    <div class="input-group-addon"><span class="fa fa-lock fa-lg"></span></div>
-                        <input type="password" class="form-control" id="password"
-                         formControlName="password" placeholder="Password">
-                    
-                  </div>
-            </div>
-        </div>
-        <div class="form-group row justify-content-center">
-          <!-- Disable submit button until form is valid  -->
-          <button class="btn btn-success" type="submit" [disabled]="!loginForm.valid">
-            <span class="fa fa-sign-in fa-lg"></span> Log In</button>
-        </div>
-      </form>
-    </div>
-  </div>
-
-</div>
