+++ client/app/landing-page/landing-page.component.html
@@ -1,24 +1,7 @@
 
 <section class="pt-5">
     <div class="container pt-5">
-        <div class="card">
-            <div class="row mb-2">
-                <div class="col">
-                    <!-- Authentication Navigation here -->
-                    <app-auth></app-auth>
-                </div>
-            </div>
-            <div class="card-block">
-                <div class="row mt-2 ">
-                    <div class="offset-sm-1 col-12 col-sm-10">
-                        <h3>Welcome to BudgetPlanner. How much have you saved today?</h3>
-                    </div>
-                    <div class="row pt-2">
-                    <router-outlet></router-outlet>
-                    </div>
-                </div>
-            </div>
-        </div>
+        <app-auth></app-auth>
     </div>
 </section>
